requirements.txt
eth-brownie==1.19.2
numpy==1.23.5
